requirements.txt
django
Markdown==3.1.1
Pillow
gunicorn==19.10.0
django-widget-tweaks==1.4.5
django-imagekit
six
django-storages[azure]